% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_classify.R
\name{build_ensemble_from_results}
\alias{build_ensemble_from_results}
\title{Build ensemble from top performing diverse models}
\usage{
build_ensemble_from_results(
  texts,
  labels,
  num_labels,
  cv_results,
  device,
  verbose
)
}
\description{
Build ensemble from top performing diverse models
}
\keyword{internal}
