% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{evaluate_classifier}
\alias{evaluate_classifier}
\title{Internal evaluation function}
\usage{
evaluate_classifier(model, tokenizer, texts, labels, batch_size, device)
}
\description{
Internal evaluation function
}
\keyword{internal}
