% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe_classifier.R
\name{evaluate_moe_classifier}
\alias{evaluate_moe_classifier}
\title{Internal evaluation function for MoE classifier}
\usage{
evaluate_moe_classifier(model, tokenizer, texts, labels, batch_size, device)
}
\description{
Internal evaluation function for MoE classifier
}
\keyword{internal}
