% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_classify.R
\name{generate_candidate_space}
\alias{generate_candidate_space}
\title{Generate candidate search space with hyperparameters}
\usage{
generate_candidate_space(
  meta_features,
  max_time_minutes,
  cv_folds,
  use_meta_learning
)
}
\description{
Generate candidate search space with hyperparameters
}
\keyword{internal}
