% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_models.R
\name{list_available_models}
\alias{list_available_models}
\title{List available pre-trained models}
\usage{
list_available_models()
}
\value{
Character vector of available model names
}
\description{
List available pre-trained models
}
\examples{
list_available_models()
}
