% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_classify.R
\name{perform_cv_selection}
\alias{perform_cv_selection}
\title{Perform cross-validation model selection}
\usage{
perform_cv_selection(
  texts,
  labels,
  num_labels,
  candidates,
  cv_folds,
  start_time,
  max_time_minutes,
  device,
  verbose
)
}
\description{
Perform cross-validation model selection
}
\keyword{internal}
