% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{predict.ensemble_classifier}
\alias{predict.ensemble_classifier}
\title{Predict with ensemble classifier}
\usage{
\method{predict}{ensemble_classifier}(object, data, text_col, type = "class", ...)
}
\arguments{
\item{object}{Ensemble classifier}

\item{data}{Data frame with text}

\item{text_col}{Column containing text}

\item{type}{Prediction type: "class", "prob", or "raw"}

\item{...}{Additional arguments}
}
\value{
Predictions data frame
}
\description{
Predict with ensemble classifier
}
