% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.granite_model}
\alias{print.granite_model}
\title{Print method for granite_model}
\usage{
\method{print}{granite_model}(x, ...)
}
\arguments{
\item{x}{granite_model object to print}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for granite_model
}
