% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.granite_tokenizer}
\alias{print.granite_tokenizer}
\title{Print method for granite_tokenizer}
\usage{
\method{print}{granite_tokenizer}(x, ...)
}
\arguments{
\item{x}{granite_tokenizer object to print}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for granite_tokenizer
}
