% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_s3.R
\name{save_classifier_impl}
\alias{save_classifier_impl}
\title{Internal: Save classifier implementation}
\usage{
save_classifier_impl(classifier, file)
}
\description{
Internal: Save classifier implementation
}
\keyword{internal}
