% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{select_diverse_candidates}
\alias{select_diverse_candidates}
\title{Select top K diverse candidates for ensemble}
\usage{
select_diverse_candidates(cv_results, k = 3)
}
\arguments{
\item{cv_results}{List of CV evaluation results}

\item{k}{Number of models to select}
}
\value{
Indices of selected candidates
}
\description{
Given CV results, select K best performing diverse models for ensembling.
Diversity is measured by prediction disagreement.
}
\keyword{internal}
