% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_classify.R
\name{select_strategy}
\alias{select_strategy}
\title{Select training strategy based on dataset characteristics}
\usage{
select_strategy(n_samples, num_labels, max_time_minutes, cv_folds, verbose)
}
\description{
Select training strategy based on dataset characteristics
}
\keyword{internal}
