% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_classify.R
\name{train_and_evaluate_fold}
\alias{train_and_evaluate_fold}
\title{Train and evaluate a single fold}
\usage{
train_and_evaluate_fold(
  train_texts,
  train_labels,
  val_texts,
  val_labels,
  num_labels,
  config,
  device,
  verbose
)
}
\description{
Train and evaluate a single fold
}
\keyword{internal}
