% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_classify.R
\name{train_final_model}
\alias{train_final_model}
\title{Train final model on all data}
\usage{
train_final_model(texts, labels, num_labels, config, device, verbose)
}
\description{
Train final model on all data
}
\keyword{internal}
