% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graniteR-package.R
\docType{package}
\name{graniteR-package}
\alias{graniteR}
\alias{graniteR-package}
\title{graniteR: Interface to IBM Granite Embedding Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a pipe-friendly interface to IBM's Granite embedding models via Python's transformers library. Supports sentence embeddings, text classification, and fine-tuning workflows following tidyverse conventions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/skandermulder/graniteR}
  \item Report bugs at \url{https://github.com/skandermulder/graniteR/issues}
}

}
\author{
\strong{Maintainer}: Skander Tahar Mulder \email{skandermulder@gmail.com}

}
\keyword{internal}
