% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graniteR-package.R
\docType{package}
\name{graniteR-package}
\alias{graniteR}
\alias{graniteR-package}
\title{graniteR: Interface to IBM Granite Embedding Models}
\description{
Provides a pipe-friendly interface to IBM's Granite embedding models
via Python's transformers library. Key features include:
\itemize{
  \item Generate sentence embeddings with \code{granite_embed()}
  \item Train text classifiers with \code{granite_classifier()}
  \item Fine-tune models with \code{granite_train()}
  \item Make predictions with \code{granite_predict()}
}
}
\section{Installation}{

The package requires Python dependencies. Use \code{install_granite()}
for fast setup with UV package manager.
}

\section{Getting Started}{

See the "Getting Started" vignette: \code{vignette("getting-started", package = "graniteR")}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/skandermulder/graniteR}
  \item Report bugs at \url{https://github.com/skandermulder/graniteR/issues}
}

}
\author{
\strong{Maintainer}: Skander Tahar Mulder \email{skandermulder@gmail.com}

}
